requirements.txt
fastapi
uvicorn
scikit-learn
numpy
joblib
torch
torchvision
pillow
